Refactor file structure, move calendar fetching to utils
